Changes to dropdowntoggle, homeContainer and services
The dropdowntoggle events were changed so the menu opens when hovered on rather than clicked on. The music-server service was added to keep information about the current music track that is playing in the musicplayer-footer. The application plays one of the most irritating songs I have ever heard. Nevertheless, it is playing music :)